chore: move hotbar related functions to hotbar_utils.js; move stats content into stats_content.html
