+++ frontend/src/js/main_menu_panel.js
@@ -1,10 +1,6 @@
 
 import { loadComponent } from "./utils/ui_utils.js";
-import { addEventListenerTo } from "./utils/ui_utils.js";
-import { getColor, getRandomColor } from "./utils/color_utils.js";
-import { hideTooltip, showTooltip } from "./utils/tooltip_utils.js";
-
-let hotbarItems
+import { initHotbar } from "./utils/hotbar_utils.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   loadMainMenuPanel()
@@ -15,155 +11,20 @@ export async function loadMainMenuPanel() {
     await loadComponent('components/menu/main_menu_panel.html')
 
     initHotbar()
+    // await loadMainMenuContent('components/menu/play_content.html')
   } catch (error) {
     console.error('Error loading Main Menu Panel :', error)
   }
 }
 
-function initHotbar() {
-  hotbarItems = {
-    'play': {
-      'name': 'play',
-      'button': document.getElementById('play-button'),
-      'icon': document.getElementById('play-icon'),
-      'tooltip': document.getElementById('play-tooltip'),
-      'isSelected': true,
-      'color': 'teal'
-    },
-    'stats': {
-      'name': 'stats',
-      'button': document.getElementById('stats-button'),
-      'icon': document.getElementById('stats-icon'),
-      'tooltip': document.getElementById('stats-tooltip'),
-      'isSelected': false,
-      'color': 'yellow'
-    },
-    'friends': {
-      'name': 'friends',
-      'button': document.getElementById('friends-button'),
-      'icon': document.getElementById('friends-icon'),
-      'tooltip': document.getElementById('friends-tooltip'),
-      'isSelected': false,
-      'color': 'blue'
-    },
-    'how-to-play': {
-      'name': 'how-to-play',
-      'button': document.getElementById('how-to-play-button'),
-      'icon': document.getElementById('how-to-play-icon'),
-      'tooltip': document.getElementById('how-to-play-tooltip'),
-      'isSelected': false,
-      'color': 'orange'
-    },
-    'settings': {
-      'name': 'settings',
-      'button': document.getElementById('settings-button'),
-      'icon': document.getElementById('settings-icon'),
-      'tooltip': document.getElementById('settings-tooltip'),
-      'isSelected': false,
-      'color': 'magenta'
-    }
+export async function loadMainMenuContent(path) {
+  try {
+    const response = await fetch(`/static/${path}`)
+    const html = await response.text()
+
+    document.querySelector('#main-menu-panel > .content-container').innerHTML = html;
+  } catch (error) {
+    console.error(`Error loading ${path}:`, error)
   }
-
-  initButton(hotbarItems['play'])
-  initButton(hotbarItems['stats'])
-  initButton(hotbarItems['friends'])
-  initButton(hotbarItems['how-to-play'])
-  initButton(hotbarItems['settings'])
 }
 
-// sets the selected button's isSelected variable to true and the rest is false
-// also sets resets all the colors of the non-selected buttons
-function updateButtonState(buttonName) {
-  hotbarItems[buttonName]['isSelected'] = true
-
-  Object.keys(hotbarItems).forEach(key => {
-    if (key !== buttonName) {
-      const buttonColor = hotbarItems[key]['color']
-
-      hotbarItems[key]['isSelected'] = false
-      hotbarItems[key]['button'].style.backgroundColor = '#000'
-      hotbarItems[key]['icon'].style.color = getColor(buttonColor, 500)
-
-      return
-    }
-  });
-}
-
-function updateBorderColor(buttonColor) {
-  const mainMenuPanel = document.getElementById('main-menu-panel')
-  const hotbar = document.getElementById('hotbar')
-
-  if (!mainMenuPanel) {
-    console.error("Error in updateBorderColor(): main-menu-panel not found")
-    return
-  }
-  if (!hotbar) {
-    console.error("Error in updateBorderColor(): hotbar not found")
-    return
-  }
-
-  mainMenuPanel.style.border = `5px solid var(--${buttonColor}-500)`
-  hotbar.style.border = `5px solid var(--${buttonColor}-500)`
-}
-
-function initButton(hotbarItemInfo) {
-  const buttonName = hotbarItemInfo['name']
-  const button = hotbarItemInfo['button']
-  const icon = hotbarItemInfo['icon']
-  const tooltip = hotbarItemInfo['tooltip']
-  const buttonColor = hotbarItemInfo['color']
-
-  addEventListenerTo(
-    button,
-    'mouseover',
-    () => {
-      showTooltip(tooltip)
-      if (!hotbarItemInfo['isSelected']) {
-        button.style.backgroundColor = getColor(buttonColor, 700)
-        icon.style.color = getColor(buttonColor, 400)
-      }
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mouseout',
-    () => {
-      hideTooltip(tooltip)
-      if (!hotbarItemInfo['isSelected']) {
-        button.style.backgroundColor = '#000'
-        icon.style.color = getColor(buttonColor, 500)
-      } else {
-        button.style.backgroundColor = getColor(buttonColor, 500)
-        icon.style.color = getColor(buttonColor, 200)
-      }
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mousedown',
-    () => {
-      if (!hotbarItemInfo['isSelected']) {
-        updateButtonState(buttonName)
-        updateBorderColor(buttonColor)
-        button.style.backgroundColor = getColor(buttonColor, 800)
-        icon.style.color = getColor(buttonColor, 600)
-      }
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mouseup',
-    () => {
-      if (!hotbarItemInfo['isSelected']) {
-        button.style.backgroundColor = '#000'
-        icon.style.color = getColor(buttonColor, 500)
-      } else {
-        button.style.backgroundColor = getColor(buttonColor, 500)
-        icon.style.color = getColor(buttonColor, 200)
-      }
-    }
-  )
-}
